refactor: yml 파일 사용 및 클래스 통합
