--- src/main/resources/connection-context.properties
@@ -1,6 +1,0 @@
-SERVER=localhost:13306
-DATABASE=chess
-DATABASE_TEST=chess_test
-OPTION=?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC
-USERNAME=root
-PASSWORD=root
