+++ src/main/resources/test_connection_context.yml
@@ -1,0 +1,5 @@
+server: localhost:13306
+database: chess_test
+option: ?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC
+username: root
+password: root
