+++ src/test/java/chess/dao/ChessGameDaoTest.java
@@ -23,8 +23,8 @@ class ChessGameDaoTest {
     @BeforeAll
     static void openConnection() {
         try {
-            ConnectionGenerator connectionGenerator = new TestConnectionGenerator();
-            connection = connectionGenerator.getConnection();
+            ConnectionGenerator connectionGenerator = new ConnectionGenerator();
+            connection = connectionGenerator.getConnection("test");
             connection.setAutoCommit(false);
         } catch (SQLException ignored) {
         }
